Update new_nav.css
Restricting the impact of the universal selector. This selector had been making subtle changes throughout the site, which have been leading to multiple small edits and PRs. Restoring the 'baseline' margins and paddings to the site will make the site marginally more responsive (reducing the application of the style to every element on the site) and reduce the number of tiny tweaks to restore readability to pages such as https://www.gutenberg.org/browse/scores/top and https://dev.gutenberg.org/help/bibliographic_record.html

I expect this change to cause some discussion. It restricts a rule that has impacted the full site in subtle ways and has unintentionally generated styling changes since last May when it was introduced.